backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api
python-dotenv
google-genai
